--- a/.github/workflows/deploy_packages.yml
+++ b/.github/workflows/deploy_packages.yml
@@ -75,6 +75,10 @@ jobs:
         with:
           node-version: ${{ matrix.node-version }}
           registry-url: https://registry.npmjs.org/ # Needed for auth
+
+      # Setup caching to optimize task execution across runs.
+      - uses: google/wireit@setup-github-actions-caching/v2@eea3c9f0385a39e6eb4ff6a6daa273311381d436
+
       - name: yarn install
         uses: backstage/actions/yarn-install@3c138326f7fcbf253b88170c1f29bae8e975d47c # v0.6.14
         with: