--- a/.github/workflows/verify_e2e-techdocs.yml
+++ b/.github/workflows/verify_e2e-techdocs.yml
@@ -41,7 +41,7 @@ jobs:
 
       # Setup caching to optimize task execution across runs.
       - uses: google/wireit@setup-github-actions-caching/v2@eea3c9f0385a39e6eb4ff6a6daa273311381d436
-      
+
       - name: install dependencies
         run: yarn install --immutable