--- a/.github/workflows/verify_e2e-windows.yml
+++ b/.github/workflows/verify_e2e-windows.yml
@@ -56,7 +56,13 @@ jobs:
           registry-url: https://registry.npmjs.org/ # Needed for auth
 
       # Setup caching to optimize task execution across runs.
-      - uses: google/wireit@setup-github-actions-caching/v2
+      # Note: Wireit documentation uses release tags, like "google/wireit@setup-github-actions-caching/v2.0".
+      #       However, Backstage uses commit hashes to ensure reproducibility.
+      #       Because Wireit uses a repo branch for their GitHub Actions code,
+      #       1. Find the appropriate release tag at https://github.com/google/wireit/tags
+      #       2. Click on the commit hash for the release tag e.g. "eea3c9f".
+      #       3. Use the full commit hash in the `uses` below.
+      - uses: google/wireit@eea3c9f0385a39e6eb4ff6a6daa273311381d436 # setup-github-actions-caching/v2.0
 
       - name: setup python
         uses: actions/setup-python@f677139bbe7f9c59b41e40162b753c062f5d49a3 # v5.2.0