--- a/.github/workflows/verify_e2e-windows.yml
+++ b/.github/workflows/verify_e2e-windows.yml
@@ -91,7 +91,7 @@ jobs:
         with:
           cache-prefix: ${{ runner.os }}-v${{ matrix.node-version }}
 
-      - run: yarn tsc
+      - run: yarn tsc:simple
       - run: yarn backstage-cli repo build
       - name: run E2E test
         run: yarn e2e-test run