--- a/.github/workflows/verify_microsite.yml
+++ b/.github/workflows/verify_microsite.yml
@@ -40,7 +40,7 @@ jobs:
           python-version: '3.9'
 
       # Setup caching to optimize task execution across runs.
-      - uses: google/wireit@setup-github-actions-caching/v2@eea3c9f0385a39e6eb4ff6a6daa273311381d436
+      - uses: google/wireit@setup-github-actions-caching/v2
 
       - name: Install MkDocs dependencies
         run: pip install mkdocs mkdocs-techdocs-core