--- a/.prettierignore
+++ b/.prettierignore
@@ -1,3 +1,5 @@
+.tasks
+.wireit
 .yarn
 dist
 microsite