--- a/CONTRIBUTING.md
+++ b/CONTRIBUTING.md
@@ -152,6 +152,10 @@ yarn new # Create a new module
 
 > See [package.json](https://github.com/backstage/backstage/blob/master/package.json) for other yarn commands/options.
 
+### Task execution caching
+
+The [Wireit](https://github.com/google/wireit) project has been introduced to optimize task execution with local caching that can also be saved and loaded across continuous integration tasks and runs. For example, `yarn tsc` is wrapped with wireit to only clean and execute when specific source files have changed. This is also a "Green IT" approach to reducing overall compute consumption for every machine that interacts with this project.
+
 ## Local configuration
 
 Backstage allows you to specify the configuration used while running the application on your computer. Local configuration is read from `app-config.local.yaml`. This file is ignored by Git, which means that you can safely use it to reference secrets like GitHub tokens without worrying about these secrets, inadvertently ending up in the Git repository. You do not need to copy everything from the default config to the local config. The `app-config.local.yaml` file will be merged with `app-config.yaml` and overwrite the default app configs.