Begin to optimize task execution across all machines with task caching using [WireIt](https://github.com/google/wireit).

This includes local task execution optimization as well as Continuous Integration with GitHub Actions caching.

## Example: prettier

_**58.94** times faster on every run!_ This is a force multiplier for developers and across CI runs.

**After:** Notice how the CI "prettier" step ran in only 1.8s and was skipped because it was cached and nothing changed in the ["CI - Verify 20.x" workflow](https://github.com/backstage/backstage/actions/runs/10757538095/job/29831704588?pr=26545#logs).

```
Run yarn prettier:check
Analyzing
  0% [0 / 1] [0 running]
✅ Ran 0 scripts and skipped 1 in 1.8s.
```

**Before:** The same "prettier" step [took 106.1s the first time](https://github.com/backstage/backstage/actions/runs/10757414687/job/29831454745#logs)!

```
Run yarn prettier:check
Analyzing
  0% [0 / 1] [0 running]
  0% [0 / 1] [1 running] prettier:check
Checking formatting...
All matched files use Prettier code style!
✅ Ran 1 script and skipped 0 in 106.1s.
```

## Example: tsc

_**19.38** times faster!_ This is a force multiplier for developers and across CI runs.

**After:** Notice how the CI "Run yarn tsc:simple" step ran in 3.7s and was skipped because it was cached and nothing changed in the ["E2E Linux 20.x" workflow](https://github.com/backstage/backstage/actions/runs/10757538098/job/29831681965?pr=26545#logs).

```text
Run yarn tsc:simple
Analyzing
  0% [0 / 1] [0 running]
✅ Ran 0 scripts and skipped 1 in 3.7s.
```

**Before:** The same "prettier" step [took 71.7s the first time](https://github.com/backstage/backstage/actions/runs/10757414674/job/29831420739#logs)!

```
1m 13s
Run yarn tsc:simple
Analyzing
  0% [0 / 1] [0 running]
  0% [0 / 1] [1 running] tsc:simple
100% [1 / 1] [0 running]
✅ Ran 1 script and skipped 0 in 71.7s.
```

## Optimized tasks

- [x] `yarn tsc:full`
- [x] `yarn tsc:simple` was introduced as a backwards-compatible method to avoid overriding expected behavior of `yarn tsc ...` command with additional arguments. For example, see ["./packages/repo-tools/src/commands/api-reports/generateTypeDeclarations.ts"](https://github.com/backstage/backstage/blob/6dca65f95b398c809c23fc5dc5e24af63febe6b8/packages/repo-tools/src/commands/api-reports/generateTypeDeclarations.ts#L34-L40)
- [x] `yarn prettier:check` required a wrapper script to send command output to a log file that can be cached.

`#GreenIT`

Signed-off-by: Eric Swanson <ericis@users.noreply.github.com>