@ericis 2024-09-08
@camilaibs , @freben , et al: Please help to re-run ["CI / Test 18.x"](https://github.com/backstage/backstage/actions/runs/10757538095/job/29831704714?pr=26545) as a test seems to have timed out unrelated to any of the changes made:

```text
FAIL plugins/search-backend-module-pg/src/database/DatabaseDocumentStore.test.ts (21.224 s)
  ● DatabaseDocumentStore › unsupported › should return support state, "MYSQL_8"

    connect ETIMEDOUT

      at Connection._handleTimeoutError (../../../node_modules/mysql2/lib/connection.js:205:17)
```