@ourchitectureio 2024-09-08
Let me know if the sort script is useful. If it is, it should probably be a Node.js script to align to the rest of the project. I have no problem deleting or rewriting.