@ourchitectureio 2024-09-25
Closing as low impact and incomplete